feat: add FreeRTOS platform support for STM32xx, including threading, timer, mutex, and semaphore implementations; update build scripts and CMake integration for RTOS selection

+++ emby/core/EmbyThreading/Thread.hh
@@ -25,7 +25,7 @@ namespace EmbyThreading
 
         Thread(EmbyThreading::Worker *worker,
                char const *name,
-               size_t stackSize,
+               uint32_t stackSize,
                Priority priority, bool start = true,
                uint32_t *stack = nullptr);
