+++ emby/platforms/Mbed/EmbyThreading/Thread.cc
@@ -30,7 +30,7 @@ static osPriority getPrio(EmbyThreading::Thread::Priority priority)
 
 EmbyThreading::Thread::Thread(EmbyThreading::Worker* worker,
                               char const*            name,
-                              size_t                 stackSize,
+                              uint32_t                 stackSize,
                               Priority               priority,
                               bool                   start,
                               uint32_t*              stack)
