+++ emby/platforms/X86_Unix/EmbyThreading/Thread.cc
@@ -13,7 +13,7 @@ namespace EmbyThreading
 
     Thread::Thread( Worker* worker,
                     char const* name,
-                    size_t stackSize,
+                    uint32_t stackSize,
                     Priority  priority , bool start,
                     uint32_t * stack )
 	{
